pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.160.0",
    "google-auth>=2.38.0",
    "google-auth-oauthlib>=1.2.1",
    "gspread>=6.2.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.37",
    "trafilatura>=2.0.0",
]
